MLOps/telco-churn-ml/requirements.txt
pandas
numpy
scikit-learn
mlflow
matplotlib
seaborn
jupyter
ipykernel
fastapi
uvicorn
joblib
pyyaml